ASP: add tooltip to remove package button in configuration windows
Rationale: we're not explaining anywhere what removing the package from
the list really means. "Removing" means "stop installing automatically"
(but not "deleting from persistence" or "uninstalling immediately", for
example). Let's keep the phrasing that we already have in the
confirmation dialog and be a bit more explicit in the tooltip, who only
people who will hesitate will see.

Will-fix: #15578