Test suite: use hopefully more reliable public GnuPG key (refs: #15771)
So far we were using anonym's key, on the ground that he would maintain
it and worst case, if he fails to keep it up-to-date, notice if automated tests
start failing. This did not work: his key has expired a couple days ago.

So let's switch to dkg's key and hope he'll maintain it better :)

While I'm at it, let's use the full fingerprint instead of a short ID when
feasible (dkg actually has published another key with the same short ID, most
likely for demonstration purposes). But Seahorse does not support fetching keys
by fingerprint so the corresponding steps keep using the short ID and I'm
updating the expected picture to match the non-revoked one of these two keys.