Delete supposedly temporary config file we instructed users to create.
On #15692 we instruct users to create preferences/0000tails.js, with the intent
that it would not be persistent because until 3.8 we have code that deletes that
file on persistence activation. But as part of the fix for #15693, I've removed
the code that deletes that file on persistence activation, which was not
necessary to fix #15693.

So a user who follows these instructions during their last Tails 3.8 session
with persistence enabled (before upgrading to 3.9) will have that file left in
persistence once they upgrade to 3.9… and then they'll be stuck forever with an
obsolete value for extensions.enigmail.configuredVersion. To avoid that, I'm
hereby re-adding the code that deletes this obsolete file.