Don't attempt to include "up-to-date" uBlock filter lists (refs: #15702).
First, since the transition to webext (commit
106a5129db7a1ff2e8415cff4d4bd88f9054ae11) we actually did not store these lists:
we merely shipped custom settings, which include the list of enabled
filter lists.

Second, even if one manually refreshes these lists, Tor Browser's "avoid disk"
feature prevents the uBlock webext from actually saving the updated lists
on "disk". So extracting them and adding them to our Git repo is not easy.

Third, given the lifetime of our ISO, any list we ship in there will always be
only marginally more up-to-date than the ones included in the Debian package.

Finally, one reason why we needed to include these lists in the ISO is that we
were enabling a couple non-default, regional ones. It makes little sense in 2018
to special-case German and French in Tails so I think this reason does not
work anymore.

So we'll now rely purely on the set of filter lists shipped in the
Debian package.