Thunderbird: check the lang packs are suffixed @firefox.mozilla.org (refs: #15091).
It was slightly surprising to spot @firefox.mozilla.org as a suffix
instead of the traditional @thunderbird.mozilla.org one. This was
reported to Debian#906176 (https://bugs.debian.org/906176).

Let's check how many language packages are found with the suffix
used in tails-shell-library/thunderbird.sh, to make sure l10n has
a chance to work.