Revert "tails-debugging-info: fix the (minimal) check for symlink attack."
This reverts commit 3460524dfc3496f0d5928c663bd0fcd646fb3aae.

As explained on https://labs.riseup.net/code/issues/8514#note-58 the commit I'm
reverting was based on a misunderstanding. While it improves things in some
cases, it makes them worse in some other cases. So let's revert it for now
until we reach an agreement on what the test should look like.