Revert "ASP: remove code to exit if running from live build"
This reverts commit c3631c8b9a852ad44a7dd9a0ce75bdca2fd440de.

Today the build fails for me with:

 The following NEW packages will be installed:
   squashfs-tools
 0 upgraded, 1 newly installed, 0 to remove and 0 not upgraded.
 Need to get 125 kB of archives.
 After this operation, 305 kB of additional disk space will be used.
 Get:1 http://deb.tails.boum.org devel/main amd64 squashfs-tools amd64 1:4.3-3.0tails4 [125 kB]
 Traceback (most recent call last):
   File "/usr/local/sbin/tails-additional-software", line 16, in <module>
     from tailslib import LIVE_USERNAME
 ImportError: cannot import name 'LIVE_USERNAME'
 Fetched 125 kB in 0s (3463 kB/s)
 E: Sub-process /usr/local/sbin/tails-additional-software apt-pre returned an error code (1)
 E: Failure running script /usr/local/sbin/tails-additional-software apt-pre

I'm not sure why it did not fail recently but in any case, I believe the
justification for commit:c3631c8b9a852ad44a7dd9a0ce75bdca2fd440de
("the ASP dpkg hooks are installed in the end of the build, making this code
useless") was based on a misunderstanding:
config/chroot_local-hooks/99-zz-install-ASP-DPKG-hooks is not "the end of the
build" and we do run APT again later. So let's reintroduce the safeguard.