VeraCrypt test suite: give up on unmounting/locking via GNOME Files.
It is way too fragile. Nautilus will sometimes fire up a "Unable to eject 105 MB
Volume" "Cannot eject drive in use: Encrypted device /dev/sda is unlocked" error
dialog while according to the Journal, everything is correctly unmounted and
locked by udisksd. There is probably a bug in Nautilus, GVfs, udisks, or the way
they talk to each other. I've occasionally seen it for years with LUKS devices
too, so this is not specific to VeraCrypt and outside of the scope of
this project.

So let's use udisksctl to unmount and lock VeraCrypt volumes instead. And as
a consequence, no notification is displayed so instead of looking for one, let's
verify that the volume was correctly unmounted (which we did already) and
locked (which we did not do, we were believing the desktop notification).