VeraCrypt test suite (file container): robustness improvement (refs: #14471, #15239).
When unlocking file containers, sometimes Dogtail would not find
the "Mount selected partition" button. This is not surprising
as I generally cannot see it when performing the exact same actions
by hand, not sure why it actually works when unlocking partitions.

Anyway, the logical thing to do is to focus the "Filesystem\n107 MB FAT"
item. It is not accessible to Dogtail so let's navigate to it with
the keyboard, which avoids having to introduce another picture
(that would need to vary depending on basic/hidden volumes because their
filesystem size differ).