Test suite: add some debugging output.
It seems to me that when a fuzzy candidate is found, instead of saving it and
aborting the scenario, our test suite goes on. In various cases this breaks the
next steps because we've not actually waited for the expected image to appear.
Let's log something when we're in the situation that might trigger this bug, in
order to make it easier to debug in the future.