Revert "Thunderbird: Import locale selection handling from Tor Browser (refs: #15091)." (refs: #15959).
This reverts commit 01df272943269352358aa1974cb6d90ddb36e926.

By merging thunderbird 1:60.0-3 from Debian, we should get a proper
setting for intl.locale.requested, which is an empty string. This
replaces the old “intl.locale.matchOs = true” behaviour.

This also makes our efforts to set this parameter based on the locale
settings moot, so let's revert it all and let the built-in code do its
job instead.