Drop useless and buggy special-casing of the t-p-s user.
Commit 3682fde35080381eb97d649966b584fc9c777998 got this wrong in two ways:

1. The goal was to *free* the UID+GID that *will* be used by the t-p-s user
   once 05-adduser_tails-persistence-setup creates it; that commit
   instead attempts to change the UID+GID of that not-yet-existing user.

2. typo ("tails-persistent-setup" instead of "tails-persistence-setup")

So this code has been a no-op for a while, which shows we don't need it.
If we ever need again to free t-p-s's UID (115) or GID (122), then we
can do that with the generic mechanism we now use for every other user & group,
i.e. by moving the "stealer" to some other UID/GID.