Use "box-sizing: border-box" everywhere
By using border-box, the border and padding of elements are including in
their width calculation.

It's what Bootstrap does by default and we've been need this trick in
several places already.

So let's embrace what Bootstrap does and adjust local.css for that.
We'll probably use Bootstrap more and more so it makes sense to align
with what they do.

For example, until now the actual width of our pages was
986px which was 930px + 2em × 2 of padding. With border-box, the width
of our pages doesn't depend anymore on the padding of its main elements.