Avoid the Trusted Reproducer getting used to trust input from the Release Manager (refs: #12629).
The previous text suggested, in a subliminal way perhaps, that the TR should
take into account, or rely on, that email, which defeated the purpose of having
a process that is not triggered by the RM. I know that we explicitly ask the TR
to do it "from [their] own initiative" but once TRs get used to receiving an
email, I suspect some of them (with suboptimal todo list management) will start
silently relying on it. Let's avoid that.