Have the TR use the .packages file generated locally instead of one provided by the RM (refs: #12169)
Since I've removed the email from the RM to the TR, it may be that the TR
did not receive that file at all. And anyway, better use one generated
from source locally instead of one provided by the RM, which could allow
a compromised RM system to perform a bunch of attacks.