Reverts commit 340ad2308267c10271ea4788ef00ece10b06e5e6.
The abovementioned commit has been made using a webinterface.
It mentioned correctly "reverting commit 7ff84722e2667326bde82a7522eade75d736bfcc".
But it is not authenticated and it looks to me that it deletes important
information. I am this restoring it. Please use Git to authenticate
yourself if you make such an important change.