Simplify
This section is meant for expert users to debug issues on Debian sid or
Debian Live. We're already not giving them much information about how to
get Debian sid or Debian Live, so it's consistent to only give them
hints about how to install Debian Live on a USB stick.