Test suite: check for "Upgrading the system" and adjust to "Upgrade successfully downloaded" new UI (#14754, #15282)
First, let's test that the "Upgrading the system" dialog is displayed while
applying an incremental upgrade. This test would have helped us notice
earlier (instead of months, if not years, too late) that this dialog was not
displayed anymore. Hopefully Sikuli is fast enough to catch it reliably: the IUK
we're applying in the test suite is pretty small so this dialog is displayed
only for a very short time.

Second, take into account that we now prompt the user before applying
the downloaded IUK. Based on work done by anonym <anonym@riseup.net>
in commit:6e6798b5d037ac41012b08f786c488884c19c88a