When the boot device is too small, wait for user input then shut down the computer (refs: #15319)
Rebooting is not very useful: in many cases, the user will hit exactly
the same problem again. Let's instead shut down the computer so
they have time to unplug the faulty USB stick.

Also, let's wait for user input instead of 10s: the user might be brewing coffee
or something and we're not in any hurry to shut down the system.