Relax check for filesystem size (refs: #16003)
https://jenkins.tails.boum.org/view/Tails_ISO/job/test_Tails_ISO_test-16003-growing-system-partition/3/cucumberTestReport/
failed with "FAT filesystem is too small: 4286566400 is less than 4294967296".

I guess that the visible size of the filesystem is smaller than the
partition that contains it because of FS data structures overhead.

In any case, what we want to check here is essentially whether the filesystem
was grown at boot time, so checking that it's at least 4000M is good enough.