Release process: automatically generate wiki/src/inc/*.html (refs: #15999)
Previously, some of these files were automatically generated (some in this
section plus some later on in the release process) but for some reason, we've
left updating the rest (manually) to the release manager. This mostly worked,
except when it didn't. Now that we need to maintain almost twice as many files
in there, so let's automate it, eventually.

Closes: #16163