Allow passing an APT snapshots config directory to "apt-snapshots-serials cat-json" (refs: #15107)
We'll need to call it on tmp/APT_snapshots.d in order to save
the *resolved* version of the serials we're going to use,
i.e. with "latest" replaced by the current latest serial.

Previously one could do that by using apt-snapshots-serials-cat-json
directly, bypassing the apt-snapshots-serials abstraction layer,
which feels wrong: I prefer apt-snapshots-serials-cat-json to
remain an implementation detail and not a stable API.