Document another workaround for Intel graphics cards (refs: #16224)
For some graphics cards, most notably 8086:0046, we force the Intel X.Org driver
(config/chroot_local-includes/usr/share/live/config/xserver-xorg/intel.ids) so
adding only nomodeset is not sufficient: the Intel driver won't work in this
context. So for those, add nomodeset and force X.Org to use the vesa driver,
which it would have automatically picked if we were not forcing the Intel one.