release_process: make it possible to set TMPDIR (refs: #16245)
The loop generating IUK files can be a bit space consuming, and it isn't
great to get ENOSPC at this stage: IUK files are prepared in place and
there's no way to tell which ones got successfully created.

Make it possible to set TMPDIR (and advertise it) to lower the chances
of running into disk space issues.