Disable live-tools.service (Closes: #16324)
This service is only useful to display the "Please remove the live-medium, close
the tray (if any) and press ENTER to continue:" prompt on shutdown, that we
don't want to display in Tails: shutdown and memory erasure should not require
a confirmation once the user has triggered it. In Stretch this code was broken
and we were relying on this. But the Buster upgrade of this code has repaired
it, so I sometimes see that prompt. This might also explain some issues such
as #16312.