Really wrap Totem with torsocks when it's started as a D-Bus service.
The fix from 24f2ccd did not do anything (anymore?) as the Exec= line
in this D-Bus service file passes the --gapplication-service option,
so the regexp would not match. As a result, Totem started from the
Applications menu was not torified, and could not open web URLs.
This was not spotted earlier because our test suite runs Totem
from a terminal, bypassing any .desktop of D-Bus .service file.