Test suite: make usb-install-tails-greeter use the Tails USB image (refs: #16004).
Let's optimize the test suite by installing the Tails USB image directly,
instead of doing the cloning dance. Remove the parent snapshot accordingly,
as there's no need to be running Tails in the first place anymore.

The Tails USB image could be trusted not to have any persistence
partition but let's keep that check anyway; that's part of a checkpoint.

The (absence of) persistence check used to appear twice; get rid of the
first instance as that one was performed after cloning, with a computer
that was still running. Keep the second one only, which happens after
the just installed system has booted.