Test suite: make installing old version of Tails quicker (refs: #16004).
It's way quicker to copy the old version of Tails using its USB image
instead of starting it from DVD and cloning to USB.

Use the implementation of the usb-install-tails-greeter checkpoint, with
a few variations (old Tails USB image, and different device name).

Also, instead of removing the USB device at the end, trigger a proper
shutdown of the computer (the emergency shutdown would be triggered
anyway).