Test suite: when MAC spoofing fails, test safety-critical properties even if the notification is not displayed.
As said on #10774, it's not clear to me whether these scenarios'
frequent failure is "merely" (sic) a test suite robustness issue,
or an actual bug in Tails.

But anyway, MAC spoofing fails, what matters above all is that the network
interfaces are disabled and that the real MAC address is not leaked, so let's
test this first: this way, even if there's a problem wrt. displaying the
notification or wrt. the test suite catching it, by running these scenarios we
will at least ensure the user is safe — possibly with poor UX, but safe.

refs: #10774