Revert "Disable snapshots entirely by never saying any of them (#16316)."
This reverts commit d82b107be9e25ba9a401dc15716c6a1f97a4b2d4.

(Sorry for the typo: the commit message was meant to be about not
*saving* snapshots…)

Initial issues encountered while smoke-testing the test suite against
buster seem to have been due to a long-standing issue in qemu (#12142
in Tails, aka. https://bugs.debian.org/851694).

That particular known issue is getting:
 - documentation: 7a6bddabe047a8ba9fb554fc3442d4fa1260e8e1
 - detection: #16355

A full test suite run with a set of patched qemu packages seems to be
just fine, so let's reinstate the snapshot mechanism.