Test suite: actually disable tails-additional-software-install.service (refs: #14596)
The system-wide tails-additional-software-install.service is not "enabled" in
Tails: instead it's started by the corresponding user
unit (/usr/lib/systemd/user/tails-additional-software-install.service).
So the previous call to "systemctl disable" was a no-op.
Let's instead disable the user service.

And then there's nothing left for us to (re-)enable: we can as well
just start the system-wide service ourselves.