Test suite: update outdated picture (refs: #14596)
The phrasing of the GUI has changed in t-p-s.git's commit
41252b861b2d3366a794d8d22808a067f9c1f19e, last April 27.

Since then this test has kept passing only because of "Found fuzzy candidate
picture for ASPPersistenceSetupOptionEnabled with similarity 0.8", which is
actually a bug (fuzzy matching is supposed to be disabled by default in our test
suite but for some reason, at least in some cases, it's enabled when it should
not).

So let's replace this picture with a current one.