Test suite: fix erroneous inter-scenario dependency doc (refs: #14596)
"I set up Additional Software when installing a package without persistent
partition and the package is installed next time I start Tails" is not
sufficient here: it does not add cowsay to the list of Additional Software.
So let's instead declare a dependency on the only scenario that
leaves cowsay in this list once it completes.