Test suite: don't look for an unrelated notification when checking if tails-additional-software-upgrade.service has started (refs: #14596)
This service does not display notifications on success: only
tails-additional-software-install.service does.

Context: I'm trying to wrap my mind around how Additional Software notifications
are handled in the test suite and it confused me a bit that we were looking for
a notification that is unrelated to what this step is testing.