Limit the memory used by mksquashfs to 512M (refs: #16177)
By default mksquashfs will use 25% of the physical memory. So when we use the
"ram" build option, build in a VM with 13GB of RAM, of which up to 12G is
supposed to be used by the build tmpfs, mksquashfs will try using 13/4 = 3.25G
of memory. And then it will get reaped by the OOM killer more or less
occasionally depending on how much space is really used in the build tmpfs and
how much memory the rest of the system is using. So let's limit the memory
used by mksquashfs to 50% of the memory we allocate to the build VM,
excluding the part of it that we expect tmpfs data to fill.

In passing, the fact mksquashfs does not get killed every time suggests that our
current BUILD_SPACE_REQUIREMENT value exceeds the real needs of a build: a value
around 10 or 11G should be enough. But that will be for another commit.