Vagrant: use xz with default settings to compress non-release SquashFS (refs: #16177)
squashfs-tools 1:4.3-11, used to build feature/buster, does not consistently
honor the value passed to -mem: the xz compressor does but at least the gzip and
lzo ones don't. This makes the build often fail because mksquashfs gets reaped
by the OOM-killer. Our only other option is currently to bump the build VM
memory a lot, which is going to be painful on developers' systems and might
not be an option on Jenkins.

So let's fall back to xz with default settings (not the crazy slow but
efficient we use at release time) when building non-release images.