Build system: rename "gzipcomp" to "fastcomp".
What matters in the "user" interface is not the exact algorithm that's used,
it's the fact it's supposed to be faster than the compression settings we use to
build releases. We may have to changes these fast(er) settings occasionally,
possibly to use a non-gzip algorithm.

So let's keep supporting "gzipcomp" for backward compatibility but stop
documenting it. Instead, support and document "fastcomp".