Test suite: only check that persistence.conf.bak exists on recent enough Tails (refs: #16461).
This makes our test suite pass even when --old-iso points to Tails 3.12, by
skipping a new test that would not work there. More specifically, the "Writing
files to a read/write-enabled persistent partition with the old Tails USB
installation" scenario would fail.

Else, whenever --old-iso is not passed at all, or it points to an ISO built from
a branch that has the code which creates persistence.conf.bak (#16461), we'll
check for its existence and access rights.