Test suite: replace workaround for bug that's been fixed with proper implementation (refs: #14596)
commit:d2a67e7f7bcafdd22186a91f3b076a01797d1107 added the workaround
and documented the bug but no ticket was filed back then. Thankfully
anonym's review highlighted this problem as a probable bug in the code,
which turned out to be the case; finally, the bug (#16475) was fixed.