Be more blurry in terms of geographical references
Some UX practitioners have been criticized to put too much emphasis on the
demographics of personas, leading to bias withing the team and sometimes
actually less diversity when it comes to the persona's goals, motivations, and
way of thinking.

See: https://medium.com/@indiyoung/describing-personas-af992e3fc527

To avoid such mistakes, we built our personas around their primary
goals. See "User goals for using Tails" in
https://tails.boum.org/contribute/personas and the process that lead to
this list.

On the other hand, personas are tools to generate empathy; giving them a
face makes them come to life. They also have to seem realistic to be
taken seriously; and sometimes that requires geographical context.

Referring to a region instead of a country, seems like a good
trade-off between being too specific and being too generic.

For example, someone could advocate for Riou to be "more political",
maybe doing direct action or something while, put in their context in
Vietnam, Riou is taking crazy risks and facing censorship like we don't
in Europe. Someone could say that the risks that we put on Cris' are too
extreme and unrealistic, when actually it's not if you know that Mexico
is on the most dangerous place for journalists: https://en.wikipedia.org
/wiki/List_of_journalists_and_media_workers_killed_in_Mexico.

Freedom of speech in Mexico and Vietnam suffer from very different
threats and Cris and Riou couldn't swap countries for example.