Release process: explain both the general case and exceptional ones wrt. UDF generation (refs: #16405).
commit:625ceaf03f6fc93e077acb1c95ce2cb6fa9871b9 automated the general
case, which is super cool; but it also:

 - Removed all the (arguably very bad) instructions we had wrt. corner cases.

   Without any explanation whatsoever, most RMs won't be able to guess whether
   they're in one of these corner cases; I bet they'll always assume they're in
   the general case and run the example command as-is (while previously they
   would have asked someone more familiar with our incremental upgrade system
   what they should do).

 - Left in place an explanation that did not match the code anymore.

This commit adds explanations of what we're trying to achieve (from the user's
PoV), how that's expressed in the example command, what the corner cases are and
how to deal with them. Hopefully these explanations will be less mind-boggling
than the previous version we had; even if they probably have plenty of problems,
I hope they'll be a good enough basis that we can improve as needs arise.