Phrase the second task to be automated in the same way as the first one (refs: #16546)
That is:

 - Describe the view with less low-level details.

 - Pretend that the automated notification system is already set up,
   linking to the ticket, so this text is super easy to update once
   it's actually the case.

Also, update phrasing to match the updated ticket gardeners' mission here: don't
suggest the ticket gardeners will ensure the review happens or the requested
info is provided. There's no way we can do that and pinging contributors on
a great number of tickets is not helpful, which is why we want a global
per-contributor notification system.