Custom APT repo: drop obsolete instruction (refs: #16576).
The version of reprepro we currently run does not choke on .buildinfo files.
The only remaining issue is that reprepro leaves these files in the "incoming"
directory after successfully importing a package, but that should be dealt with
elsewhere (e.g. on the infra side) — if at all — instead of requiring
everyone who uploads a package to do some extra work.