Make verification process more straightforward (refs: #12629)
The previous instructions did the check we want… *transitively*: we checked that
the published products matched SHA512SUMS.txt, in a step that claims to
compare them with locally built products; this was only valid because
of the implicit (and so far correct) assumption that, at this point,
we've already verified that what you have built matches SHA512SUMS.txt.

This adds complexity and makes it needlessly hard to understand and review
this process. Instead, let's directly do the check we want.