Directly verify what we mean to (refs: #12629)
When I automated this step, I was confused due to the problem I've fixed in the
previous commit, so I wrote code that checks the published products against the
IDF, while here we want to check the locally built products against the IDF.

Granted, here again, this was valid *transitively*, because of the implicit (and
so far correct) assumption that, at this point, we've already verified that what
we have built matches SHA512SUMS.txt, and that the published products match
SHA512SUMS.txt too, so both sets of products are equivalent at this point.

Still, for the same reason as in the previous commit, let's simplify and
directly do the check we want.