Release process: remove confusing and mostly obsolete instructions (Closes: #16598)
This sanity check was added after a RM or three made mistakes in the "Prepare
upgrade-description files" section, which was recently improved, so one
may hope that this check is not needed anymore.

kibi didn't understand at all these instructions, which suggests
they are not useful.

And anyway, worst case, if a UDF is missing, one of our automated test cases
will fail, which should alert us about the problem.