Release process: avoid having to guess when it's not needed (refs: #16600)
kibi wrote "one is expected to know about the next major release, and about the
one after that". I assume that's about $SECOND_NEXT_PLANNED_MAJOR_VERSION.
In the general case, indeed we don't always know what that version will be.

Thankfully, we only need this variable when preparing a RC for a major release,
in which case all the RM needs to know is 1. the version number of that major
release (the RM has to know that anyway); and 2. the version number of the major
release after the one the they're preparing (in general we have a release
schedule for the next 6-12 months, which includes this info).

So let's only set this variable when we need it, i.e. when it's rather easy
to find out what its value should be: as kibi said, right now we have
no idea what will be the major release after 3.17.