Release process: clarify what does not need to be adapted (refs: #16600)
Given we're going to have a bunch of bugfix releases in a row, recent RMs have
been tempted to set $NEXT_PLANNED_MAJOR_VERSION to one of them, assuming that if
they didn't do so, the rest of our release process doc would be buggy because
said doc assumes strictly alternating bugfix/major releases. AFAIK this is not
the case: to the best of my knowledge, the release process doc makes no such
assumption anymore (it used to, but we've been in such a situation already
and had to make the doc support it).

So clarify that this part of the release process doc can be followed blindly.