Disable the TopIcons GNOME Shell extension (refs: #16608)
This extension causes crashes (#11188).

It does not work on Wayland (#8309, #12213) so long-term, we need to remove
it anyway.

So far we've been blocking on the WIP conversion of OpenPGP Applet into a proper
GNOME Shell extension (#8310) but it's no clear whether it's worth the effort.

In order to learn how much our users rely on this extension and — transitively —
on OpenPGP Applet, let's disable this extension for one Tails release. While
TopIcons is disabled by default:

 - Users can still use OpenPGP Applet via the system tray in the bottom
   left corner of the desktop.
 - Users who do need TopIcons for other reasons can enable it again
   with 1 command line.