Test suite: disable tests about notifications in case of MAC spoofing failure (refs: #10774)
These notifications are not always displayed, which seems to be a real bug
and not a test suite robustness issue. This bug has been here for more
than three years. In the last 5 full test suite runs on the stable branch,
these steps failed 5/3×5 = 40% of the time. This does not give us any
useful information, other than confirming the bug, so let's drop these steps,
which incidentally allows us to remove the @fragile tags on the corresponding
scenarios ⇒ greater chances we detect issues in the "fail closed" property,
which is great.