Install Electrum from sid (refs: #16204).
Let's make feature/buster closer to 3.x and to what we would like Tails 4.0 to
be; we have a few months left until the 4.0 release to define our strategy wrt.
Electrum, which is being discussed on tails-dev@, and so far it's unlikely that
we'll remove it from Tails; it's likely that we'll end up installing the package
from sid (if feasible) or shipping a custom backport for Buster (otherwise), and
the package currently in sid is the closest thing available at the moment.

This reverts commit e192c5d27afc5ddfb7fda09bd233f9024d1837fd,
282920f0e2cdf6f7483e90c8b4ce2ce8dda80cf6,
and 483c87dd82d35fa4aacf73911a34c4cbd043ca98.