Test suite: remove tests that exercise Gobby or otherwise rely on it (refs: #15291)
We lose the "Explicitly torify-wrapped applications are using the default
SocksPort" and "Explicitly torsocks-wrapped applications are using the default
SocksPort" scenarios along the way. We could of course rewrite them to test
something else than Gobby but I see very little value in it: in practice,
they merely test that the default configuration of torsocks and torify
is what it has been since these pieces of software were created.