Allow submitting code for review without waiting for Jenkins to run the entire test suite.
Let's explicitly allow submitting a branch immediately after running the
relevant test suite scenarios locally. This should make it possible for
a developer to start working on a ticket and submit the resulting branch for QA
in one single work session: while the branch builds on Jenkins, one can run the
relevant test suite scenarios locally. I hope this can lower the amount of WIP
developers carry and the number of context switches they have to do
(refs: #15864).

Worst case, we'll notice any test suite regression after the merge;
we can then revert the merge and nicely ask the branch submitter
to fix the regression before we merge it again.