Add forms to search our public mailing lists on about/contact (refs: #9904).
Since more than 3 years, #9904 has been handled as part of #9900 ("Improve
Website search"). I don't believe this approach is going anywhere because under
the current assumption that our website must not use an external search engine,
this would require ikiwiki to index the archive of our mailing lists. I don't
think that's realistic (no progress on any of the ikiwiki-related subtasks
of #9900 for years) and even if it were, I don't think it's a good use of our
time.

In the specific case of our public mailing lists, I believe we can skip the
discussion about using an external search engine: our mailing lists are already
hosted by a third party (Autistici/Inventati) and we already provide forms to
subscribe to them, which POST to that third party. So adding another set of
forms that POST to A/I's mailing list archive search engine does not make any
difference in this respect.

To improve UX, I'm using a placeholder that should help visitors of these web
pages understand what the query string should look like (before they type
anything), and a tiny bit of JavaScript that initializes the value to the
required query prefix once the visitor has selected the search field.
This requires disabling the htmlscrubber ikiwiki plugin on the affected pages.

Finally, I'm wrapping these forms with <p>…</p>, otherwise the rendered layout
is much too packed.