Test suite: make tails-security-check's SOCKS port test work when there's a live security advisory (refs: #16701)
With a live security advisory, an instance of tails-security-check is already
running, which breaks "I re-run tails-security-check" (timeout waiting for the
process to exit). Let's instead re-run it via its systemd service, so that:

1. any already running tails-security-check is killed, which fixes this bug;

2. we test tails-security-check in an environment closer to how it's run in
   a real Tails: as we can see on #16603, running this kind of scripts via
   systemd or without it can yield subtly different behavior.

But systemctl returns as soon as the process is started, while our
previous implementation waited for tails-security-check to exit.
So we need to wait a little bit before analysing the network trace.