Don't install the firmware-linux-nonfree metapackage (refs: #16708).
Its only purpose is to depend on firmware-misc-nonfree and
firmware-amd-graphics, and to recommend intel-microcode and amd64-microcode.
We explicitly install all these packages already. The dependency is versioned
which breaks our build at the moment, since we want to ship a different
version of firmware-amd-graphics (4fbcc2a1013f8eff16de5a933932634aed4da042).